fixm compilation warnings about time_t structure and lld printf

+++ trunk/parseable.c
@@ -237,7 +237,7 @@ parseout(time_t curtime, int numsecs,
 			snprintf(header, sizeof header, "%s %s %ld %s %s %d",
 				labeldef[i].label,
 				utsname.nodename,
-				curtime,
+				(long int)curtime,
 				datestr, timestr, numsecs);
 
 			/*
@@ -523,7 +523,7 @@ print_PRG(char *hp, struct sstat *ss, struct pstat *ps, int nact)
 				ps->gen.pid,	/* was TGID formerly */
 				ps->gen.nthr,
 				ps->gen.excode,
-				ps->gen.btime,
+				(long int)ps->gen.btime,
 				ps->gen.cmdline,
 				ps->gen.ppid,
 				ps->gen.nthrrun,
@@ -535,7 +535,7 @@ print_PRG(char *hp, struct sstat *ss, struct pstat *ps, int nact)
 				ps->gen.sgid,
 				ps->gen.fsuid,
 				ps->gen.fsgid,
-				ps->gen.elaps);
+				(long int)ps->gen.elaps);
 	}
 }
